:bug: fix some compiler errors

+++ src/graphql/acsys/datastream/filterdupes.rs
@@ -91,7 +91,7 @@ mod test {
         global::DataInfo {
             timestamp: ts,
             result: global::DataType::Scalar(global::Scalar {
-                scalar_value: ts / 2.0,
+                scalar_value: (ts / 2.0) as f32,
             }),
         }
     }
